:recycle: RUMM-37 Update the file I/O mecanism
When writing a log, we can only append to a file if it's less than 5 seconds old.
When reading from a log file, we can only read i the file is older than 5 seconds old.
This mecanisme ensures that we wont append a log on a file being read and risk losing data;
it also ensures we don't read from a file being written and risk sending the same log twice.