Expose Telemetry and addLongTask for internal use
We expose internal methods through the _InternalProxy class, which is documented for internal use, and the `_internal` static member on the Datadog object. We keep Telemetry and RUM as separate proxies, in case we have to split them as v2 takes shape.

Exposing Telemetry requires adding a way for reporting errors that are not Throwable (for use in cross platform frameworks), so I modified the event to take `stack` and `kind` as strings instead.