RUMM-2599 Remove a flakey test
The test to ensure that events sent before the application was displayed that orphan silently, should not use the logger was failing because any events that are sent to the ViewManagerScope when the application is not displayed are no longer orphaned, as they go to the ApplicationLaunch view. Occasionally, the test would log something unrelated to orphaning which was failing the test.