fix: Allow JankStats to be explicitly disabled via additionalConfig
refs: RUM-12763

This PR adds a check to RUM feature initialization: if the new internal
option `_dd.rum.disable_jank_stats` is explicitly set to true, then we
refrain from calling `initializeFrameStatesAggregator()`. Otherwise, we
call that function as normal.

This allows the Unity SDK to configure the Android SDK in such a way
that it will never call `JankStats.createAndTrack()`, which is necessary
to avoid crashes caused by the older versions of `androidx.metrics` that
the `dd-sdk-unity` is forced to use in older versions of Unity.

(I notice that the generated `apiSurface` files are not up to date at
latest in develop, so I expect that CI will fail. When I test locally
with `./local-ci.sh --test`, everything passes. I can rebase once the
unrelated API surface issue is fixed in develop.)